add control

+++ uart.h
@@ -5,4 +5,4 @@
 void tx_byte(byte c);
 void setup_uart(unsigned long bitrate);
 void print_int_hex(byte val);
-void print_int_dec(byte val);
+void print_int_dec(uint16_t val);
